Add event ticket sales, wire real events data everywhere
Last item from the SpotHopper gaps analysis. Scoping this surfaced
that the public events pages were entirely disconnected from the CMS
(/events was a bare redirect, /events/[slug] ignored its own slug),
and the homepage + /reservations calendar both read a hardcoded
static array instead of the real events table — both confirmed with
the user before fixing, since the second one touches the reservation
booking flow directly.

- Ticket checkout: lib/clover.ts's createTicketCheckoutSession and
  lib/stripe.ts's createTicketStripeCheckoutSession look the event up
  server-side (never trust a client price), app/api/checkout/tickets
  provider-switches on PAYMENTS_PROVIDER same as the shop, checks
  capacity with the same pending-window race guard already used for
  reservations, and short-circuits events priced at $0 (the schema
  default) straight to a paid order instead of a broken zero-dollar
  checkout.
- orders gains an event_id column; lib/orders.ts's markOrderPaid/
  markOrderFailed now RETURNING * so callers can tell what matched.
- The Clover webhook only ever fulfilled bookings before this — it now
  also fulfills orders (shop and tickets), closing a real gap where
  Clover-paid purchases were never marked paid at all. Stripe's webhook
  gets the same ticket-confirmation-email branch.
- Real events data everywhere: /events and /events/[slug] rewritten to
  read the CMS (getPublishedEvents/getEventBySlug, first real public
  use of those reads), and weekly-calendar-carousel.tsx /
  experiences-lineup.tsx now take an `events` prop instead of importing
  the static lib/data.ts array — their date-grouping/animation logic is
  untouched, only the data source changed. app/reservations/page.tsx
  split into a server shell + reservations-client.tsx since it owns
  wizard state and can't fetch server-side directly.
- Fixed a real static-prerender gap this surfaced: `/`, `/events`, and
  `/reservations` were being frozen at build time with no revalidation
  — added `revalidate = 60` so a newly published event actually shows
  up without a redeploy.
- Staff events admin gains a capacity field (blank = unlimited).

Two real bugs found by actually testing this, not by inspection: the
Neon driver returns Postgres `numeric` columns and `SUM()`/`count()`
aggregates as strings, not JS numbers. `event.price === 0` never
matched (silently skipping the free-event path), and
ticketsSoldOrHeld's `held + quantity` was doing string concatenation
('1' + 1 = '11'), which happened to still reject correctly by
coincidence in earlier manual checks but broke a boundary case in
testing. Fixed at the source in lib/clover.ts, lib/stripe.ts, and
app/api/checkout/tickets/route.ts; also proactively fixed the same
latent issue in lib/subscribers.ts's subscriberCounts (same pattern,
would have shown "5" formatted wrong instead of a real bug, caught
before it mattered).

Verified end-to-end on a real test event: rendered on all four
surfaces (events list, event detail, homepage, reservations), bought a
ticket through both Clover and Stripe, hit the capacity limit exactly
at the boundary and got correctly rejected, replayed signed webhook
events for both providers and confirmed orders flipped to paid with
confirmation emails sent, and confirmed the free-event path records an
already-paid order with no payment step. All test data cleaned up.

+++ lib/subscribers.ts
@@ -36,11 +36,16 @@ export async function listSubscribers(limit = 200, offset = 0): Promise<{ rows: 
 
 export async function subscriberCounts(): Promise<{ email: number; sms: number; total: number }> {
   const sql = getSql()
+  // count(*)::int — without the cast, Postgres returns bigint, which
+  // this driver hands back as a string, not a number (same issue fixed
+  // in lib/orders.ts's ticketsSoldOrHeld — found there by a real test
+  // that broke on it; fixing it here proactively since it's the same
+  // pattern).
   const rows = (await sql`
     SELECT
-      count(*) FILTER (WHERE email IS NOT NULL AND email_opt_in) AS email,
-      count(*) FILTER (WHERE phone IS NOT NULL AND sms_opt_in) AS sms,
-      count(*) AS total
+      count(*) FILTER (WHERE email IS NOT NULL AND email_opt_in)::int AS email,
+      count(*) FILTER (WHERE phone IS NOT NULL AND sms_opt_in)::int AS sms,
+      count(*)::int AS total
     FROM subscribers
   `) as { email: number; sms: number; total: number }[]
   return rows[0] ?? { email: 0, sms: 0, total: 0 }
